Fine tune 1

+++ backend/controllers/admin/Movie.js
@@ -6,6 +6,7 @@ import {
 } from "../../config/Cloudinary.js";
 
 //add movie
+//--
 const OMDB_API_KEY = process.env.OMDB_API_KEY;
 
 export const addMovie = async (req, res) => {
@@ -34,11 +35,10 @@ export const addMovie = async (req, res) => {
       return res.status(404).json({ message: "Movie not found in OMDb !" });
     }
 
-    //parse runtime
+    //format runtime
     const runtimeMatch = omdbData.Runtime.match(/\d+/);
     const runtime = runtimeMatch ? parseInt(runtimeMatch[0]) : 0;
 
-    //check if movie already exists
     const existingMovie = await Movie.findOne({ title: omdbData.Title });
     if (existingMovie) {
       return res.status(400).json({ message: "Movie already exists !" });
@@ -49,13 +49,13 @@ export const addMovie = async (req, res) => {
       (file) => file.path
     );
 
-    //upload files to cloudinary
+    //files upload to cloud
     const mainImageUrl = await uploadToCloudinary(mainImagePicPath);
     const galleryImagesUrls = await uploadMultipleFilesToCloudinary(
       galleryImagesPaths
     );
 
-    //save in mongodb
+    //create movie
     const movie = await Movie.create({
       title: omdbData.Title,
       status,
@@ -88,6 +88,7 @@ export const addMovie = async (req, res) => {
       }
     });
 
+    //create showtimes
     for (const element of time) {
       await Showtime.create({
         movieId: movie._id,
@@ -102,18 +103,19 @@ export const addMovie = async (req, res) => {
       movie,
     });
   } catch (error) {
-    console.error("Upload failed !:", error.message);
-    res.status(500).json({ message: "Upload failed !" });
+    console.error("Movie add failed !:", error.message);
+    res.status(500).json({ message: "Movie add failed !" });
   }
 };
 
 //update movie
+//--
 export const updateMovie = async (req, res) => {
   try {
     const { closeDate, movieId } = req.body.values;
 
     //validation
-    if (!closeDate) {
+    if (!closeDate || !movieId) {
       return res.status(400).json({ message: "All fields are required !" });
     }
 
@@ -131,16 +133,24 @@ export const updateMovie = async (req, res) => {
       movie,
     });
   } catch (error) {
-    console.error("Update failed !:", error.message);
-    res.status(500).json({ message: "Update failed !" });
+    console.error("Movie update failed !:", error.message);
+    res.status(500).json({ message: "Movie update failed !" });
   }
 };
 
 //delete movie
+//--
 export const deleteMovie = async (req, res) => {
   try {
     const { Id } = req.params;
+
+    //validation
+    if (!Id) {
+      return res.status(400).json({ message: "Id not provided !" });
+    }
+
     const movie = await Movie.findByIdAndDelete(Id);
+
     if (!movie) {
       return res.status(404).json({ message: "Movie not found !" });
     }
@@ -152,7 +162,7 @@ export const deleteMovie = async (req, res) => {
       movie,
     });
   } catch (error) {
-    console.error("Delete failed !:", error.message);
-    res.status(500).json({ message: "Delete failed !" });
+    console.error("Movie delete failed !:", error.message);
+    res.status(500).json({ message: "Movie delete failed !" });
   }
 };
